Refactor: MainPage.tsx mainpage 폴더 내부로 이동
